Feature/dir selection

+++ scripts/install_controller.sh
@@ -15,7 +15,7 @@ _K8S_MGMT_HELM_UPGRADE_COMMAND="upgrade"
 function installPersistenceVolumeClaim() {
     # variables
     local __INTERNAL_PROJECT_DIRECTORY
-    dialogAskForProjectDirectory __INTERNAL_PROJECT_DIRECTORY
+    dialogAskForProjectDirectory __INTERNAL_PROJECT_DIRECTORY "true"
     local __INTERNAL_NAMESPACE
     dialogAskForNamespace __INTERNAL_NAMESPACE
 
@@ -76,7 +76,8 @@ function checkAndInstallNamespace() {
 
     if [[ -n "${__INTERNAL_NAMESPACE_TO_CHECK}" ]]; then
         # lookup if namespace already exists
-        local __INTERNAL_NS_EXISTS=$(kubectl get namespaces | awk '{print $1}' | grep "${__INTERNAL_NAMESPACE_TO_CHECK}")
+        local __INTERNAL_NS_EXISTS
+        __INTERNAL_NS_EXISTS=$(kubectl get namespaces | awk '{print $1}' | grep "${__INTERNAL_NAMESPACE_TO_CHECK}")
 
         if [[ -z "${__INTERNAL_NS_EXISTS}" ]]; then
             kubectl create namespace "${__INTERNAL_NAMESPACE_TO_CHECK}"
@@ -111,7 +112,7 @@ function installOrUpgradeJenkins() {
     local __INTERNAL_HELM_JENKINS_PATH="./charts/jenkins-master"
     # get project directory
     local __INTERNAL_PROJECT_DIRECTORY_NAME
-    dialogAskForProjectDirectory __INTERNAL_PROJECT_DIRECTORY_NAME
+    dialogAskForProjectDirectory __INTERNAL_PROJECT_DIRECTORY_NAME "true"
     # get namespace from global variables or ask for the name
     local __INTERNAL_NAMESPACE
     dialogAskForNamespace __INTERNAL_NAMESPACE
@@ -149,7 +150,18 @@ function installOrUpgradeJenkins() {
     installPersistenceVolumeClaim
 
     # install or upgrade the Jenkins Helm Chart
-    helm "${__INTERNAL_HELM_COMMAND}" "${JENKINS_MASTER_DEPLOYMENT_NAME}" "${__INTERNAL_HELM_JENKINS_PATH}" -n "${K8S_MGMT_NAMESPACE}" -f "${__INTERNAL_FULL_PROJECT_DIRECTORY}/jenkins_helm_values.yaml"
+    if [[ -f "${__INTERNAL_FULL_PROJECT_DIRECTORY}/jenkins_helm_values.yaml" ]]; then
+        helm "${__INTERNAL_HELM_COMMAND}" "${JENKINS_MASTER_DEPLOYMENT_NAME}" "${__INTERNAL_HELM_JENKINS_PATH}" -n "${K8S_MGMT_NAMESPACE}" -f "${__INTERNAL_FULL_PROJECT_DIRECTORY}/jenkins_helm_values.yaml"
+    else
+        echo ""
+        echo "  INFO: No Jenkins Helm values found..."
+        echo ""
+    fi
+
+    # lookup for scripts to execute
+    if [[ -d "${__INTERNAL_FULL_PROJECT_DIRECTORY}/scripts/" ]]; then
+        find "${__INTERNAL_FULL_PROJECT_DIRECTORY}/scripts/" -name "i_*.sh" -type f -exec chmod +x {} \; -exec {} \;
+    fi
 }
 
 ##########
@@ -161,7 +173,7 @@ function installIngressControllerToNamespace() {
     local __INTERNAL_HELM_NGINX_INGRESS_PATH="./charts/nginx-ingress-controller"
     # get project directory
     local __INTERNAL_PROJECT_DIRECTORY_NAME
-    dialogAskForProjectDirectory __INTERNAL_PROJECT_DIRECTORY_NAME
+    dialogAskForProjectDirectory __INTERNAL_PROJECT_DIRECTORY_NAME "true"
     # get namespace from global variables or ask for the name
     local __INTERNAL_NAMESPACE
     dialogAskForNamespace __INTERNAL_NAMESPACE
@@ -169,8 +181,13 @@ function installIngressControllerToNamespace() {
     # create new variable with full project directory
     local __INTERNAL_FULL_PROJECT_DIRECTORY="${PROJECTS_BASE_DIRECTORY}${__INTERNAL_PROJECT_DIRECTORY_NAME}"
 
-    # install the nginx-ingress controller with loadbalancer and default route
-    helm install "${NGINX_INGRESS_DEPLOYMENT_NAME}" "${__INTERNAL_HELM_NGINX_INGRESS_PATH}" -n "${__INTERNAL_NAMESPACE}" -f "${__INTERNAL_FULL_PROJECT_DIRECTORY}/nginx_ingress_helm_values.yaml"
+    if [[ -f "${__INTERNAL_FULL_PROJECT_DIRECTORY}/nginx_ingress_helm_values.yaml" ]]; then
+        # install the nginx-ingress controller with loadbalancer and default route
+        helm install "${NGINX_INGRESS_DEPLOYMENT_NAME}" "${__INTERNAL_HELM_NGINX_INGRESS_PATH}" -n "${__INTERNAL_NAMESPACE}" -f "${__INTERNAL_FULL_PROJECT_DIRECTORY}/nginx_ingress_helm_values.yaml"
+    echo ""
+        echo "  INFO: No Nginx Helm values found..."
+        echo ""
+    fi
 }
 
 ##########
@@ -197,4 +214,9 @@ function uninstallJenkins() {
     dialogAskForDeploymentName __INTERNAL_HELM_DEPLOYMENT_NAME
 
     helm uninstall "${__INTERNAL_HELM_DEPLOYMENT_NAME}" -n "${__INTERNAL_NAMESPACE}"
+
+    # lookup for scripts to execute if namespace name = directory name
+    if [[ -d "${PROJECTS_BASE_DIRECTORY}${__INTERNAL_NAMESPACE}/scripts/" ]]; then
+        find "${PROJECTS_BASE_DIRECTORY}${__INTERNAL_NAMESPACE}/scripts/" -name "d_*.sh" -type f -exec chmod +x {} \; -exec {} \;
+    fi
 }
